+++ config/config.php
@@ -14,6 +14,7 @@ return [
      * Keko naming
      */
     'keko' => 'Duck',
+
     /*
      * The route the CMS core will get to build the website.
      */