+++ src/Http/Controllers/Auth/ConfirmPasswordController.php
@@ -27,8 +27,6 @@ class ConfirmPasswordController extends Controller
 
     /**
      * Create a new controller instance.
-     *
-     * @return void
      */
     public function __construct()
     {