+++ src/resources/views/fuse/layouts/head.blade.php
@@ -7,6 +7,5 @@
 @yield('css')
 
  <!-- App css -->
-<link href="{{ asset('vendor/duck-funk-core/css/bootstrap.min.css') }}" rel="stylesheet" type="text/css" />
-<link href="{{ asset('vendor/duck-funk-core/css/icons.css') }}" rel="stylesheet" type="text/css" />
+<link href="{{ asset('vendor/duck-funk-core/css/core.css') }}" rel="stylesheet" type="text/css" />
 <link href="{{ asset('vendor/duck-funk-core/css/style.css') }}" rel="stylesheet" type="text/css" />