+++ src/routes/duck-funk.php
@@ -25,7 +25,6 @@ Route::group(['middleware' => ['web', 'auth', BanMiddleware::class], 'namespace'
      * Check where we start to construct the routes based in the config route param
      */
     Route::prefix(config('duck-funk.route'))->group(function () {
-        Route::get('home', 'DuckController')->name('home');
         Route::get('expulsion', function () {
             return view('duck-funk-core::ban');
         })->name('ban');
@@ -38,10 +37,13 @@ Route::group(['middleware' => ['web', 'auth', BanMiddleware::class], 'namespace'
         Route::group(['middleware' => [HousekeepingMiddleware::class, 'password.confirm'], 'namespace' => 'Housekeeping'], function () {
             Route::prefix(config('duck-funk.housekeeping_route'))->group(function () {
                 Route::resource('news', 'NewsController');
+                Route::resource('pages', 'PageController');
                 Route::get('/', 'DashboardController@index')->name('housekeeping');
                 Route::post('dashboard-parser', 'DashboardController@getUpdateWall')->name('dashboard-parser');
             });
         });
+
+        Route::get('{slug}', ['uses' => 'PageController@getPage'])->where('slug', '([A-Za-z0-9\-\/]+)');
     });
 });